Added category information to each algorithm. Also cleaned how a singular algorithm is displayed.
